mpl: extract Pusher for better testing
