+++ src/mpl/BUILD
@@ -26,6 +26,26 @@ cc_library(
 )
 
 cc_library(
+    name = "pusher",
+    srcs = [
+        "src/object.cpp",
+        "src/pusher.cpp",
+        "src/shapes.cpp",
+    ],
+    hdrs = [
+        "src/mpl-util.h",
+        "src/object.h",
+        "src/pusher.h",
+        "src/shapes.h",
+    ],
+    deps = [
+        "//src/odb",
+        "//src/utl",
+        "@boost.random",
+    ],
+)
+
+cc_library(
     name = "mpl",
     srcs = [
         "src/MplObserver.h",
@@ -39,12 +59,7 @@ cc_library(
         "src/clusterEngine.h",
         "src/hier_rtlmp.cpp",
         "src/hier_rtlmp.h",
-        "src/mpl-util.h",
-        "src/object.cpp",
-        "src/object.h",
         "src/rtl_mp.cpp",
-        "src/shapes.cpp",
-        "src/shapes.h",
     ],
     hdrs = [
         "include/mpl/rtl_mp.h",
@@ -53,6 +68,7 @@ cc_library(
         "include",
     ],
     deps = [
+        ":pusher",
         ":snapper",
         "//src/dbSta",
         "//src/dbSta:dbNetwork",
@@ -74,9 +90,6 @@ cc_library(
         "src/clusterEngine.h",
         "src/graphics.cpp",
         "src/graphics.h",
-        "src/mpl-util.h",
-        "src/object.h",
-        "src/shapes.h",
         ":swig",
         ":tcl",
     ],
@@ -89,6 +102,7 @@ cc_library(
     ],
     deps = [
         ":mpl",
+        ":pusher",
         "//:ord",
         "//src/gui",
         "//src/odb",
