+++ src/mpl/src/pusher.h
@@ -1,0 +1,56 @@
+// SPDX-License-Identifier: BSD-3-Clause
+// Copyright (c) 2026, The OpenROAD Authors
+
+#pragma once
+
+#include <map>
+#include <vector>
+
+#include "mpl-util.h"
+#include "object.h"
+#include "odb/db.h"
+#include "odb/geom.h"
+
+namespace utl {
+class Logger;
+}  // namespace utl
+
+namespace mpl {
+
+class Pusher
+{
+ public:
+  Pusher(utl::Logger* logger,
+         Cluster* root,
+         odb::dbBlock* block,
+         const std::vector<odb::Rect>& io_blockages);
+
+  void pushMacrosToCoreBoundaries();
+
+ private:
+  void setIOBlockages(const std::vector<odb::Rect>& io_blockages);
+  bool designHasSingleCentralizedMacroArray();
+  void pushMacroClusterToCoreBoundaries(
+      Cluster* macro_cluster,
+      const std::map<Boundary, int>& boundaries_distance);
+  void fetchMacroClusters(Cluster* parent,
+                          std::vector<Cluster*>& macro_clusters);
+  std::map<Boundary, int> getDistanceToCloseBoundaries(Cluster* macro_cluster);
+  void moveHardMacro(HardMacro* hard_macro, Boundary boundary, int distance);
+  void moveMacroClusterBox(odb::Rect& cluster_box,
+                           Boundary boundary,
+                           int distance);
+  bool overlapsWithHardMacro(const odb::Rect& cluster_box, int cluster_id);
+  bool overlapsWithIOBlockage(const odb::Rect& cluster_box) const;
+
+  utl::Logger* logger_;
+
+  Cluster* root_;
+  odb::dbBlock* block_;
+  odb::Rect core_;
+
+  std::vector<odb::Rect> io_blockages_;
+  std::vector<HardMacro*> hard_macros_;
+};
+
+}  // namespace mpl
