+++ src/mpl/test/cpp/CMakeLists.txt
@@ -23,7 +23,14 @@ gtest_discover_tests(TestBlockMacroChannels
 )
 
 add_executable(TestPusher TestPusher.cpp)
-target_link_libraries(TestPusher ${TEST_LIBS})
+target_link_libraries(TestPusher
+  GTest::gtest
+  GTest::gtest_main
+  GTest::gmock
+  odb
+  mpl_pusher
+  tst
+)
 gtest_discover_tests(TestPusher WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR})
 
 add_dependencies(build_and_test
