+++ src/mpl/test/cpp/TestPusher.cpp
@@ -6,8 +6,8 @@
 #include <utility>
 #include <vector>
 
-#include "../../src/hier_rtlmp.h"
 #include "../../src/object.h"
+#include "../../src/pusher.h"
 #include "MplTest.h"
 #include "gtest/gtest.h"
 #include "odb/db.h"
